% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Matrix_Report.R
\name{build_vce_mat_report}
\alias{build_vce_mat_report}
\title{Build a matrix with VCE results for a report}
\usage{
build_vce_mat_report(
  ps_random_effect,
  ps_offdiagonal,
  ps_diagonal,
  pvec_trait_name,
  pvec_vce_results_file_path
)
}
\arguments{
\item{ps_random_effect}{random effect (for exemple: animal, herdyear, residual)}

\item{ps_offdiagonal}{covariance or genetic_correlation}

\item{ps_diagonal}{variance or heritability}

\item{pvec_vce_results_file_path}{path to csv-file with vce results}
}
\value{
mat_randomEffect resulting matrix with vce results from the random effect
}
\description{
Build a matrix with VCE results for a report
}
\details{
---
title: "Building a matrix from VCE results for the report"
date:  "2021-12-21"
#' ---
}
