% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_cor_conversion.R
\name{cor_to_cov}
\alias{cor_to_cov}
\title{Conversion of Correlation Matrix and Vector of Variances to Variance-Covariance Matrix}
\usage{
cor_to_cov(pmat_cor, pvec_var)
}
\arguments{
\item{pmat_cor}{correlation matrix}

\item{pvec_var}{vector of variance components}
}
\value{
variance covariance matrix
}
\description{
The same approach as in `base::cov2cor()` is used to compute the variance-covariance 
matrix from a matrix of correlations and a vector of variance values.
}
\examples{
## Taken from ?cov2cor: Correlation Matrix of Multivariate sample:
Cv <- cov(longley)
Cl <- cor(longley)
stopifnot(all.equal(cor_to_cov(Cl, diag(Cv)), Cv))
}
