% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_cor_conversion.R
\name{cov_to_cor}
\alias{cov_to_cor}
\title{Conversion of covariances to correlations}
\usage{
cov_to_cor(V)
}
\arguments{
\item{V}{symmetric positive definite variance-covariance matrix}
}
\value{
correlation matrix
}
\description{
Use the function `base::cov2cor()` to convert a variance-covariance 
matrix to a correlation matrix
}
\examples{
## Taken from ?cov2cor: Correlation Matrix of Multivariate sample:
(Cv <- cov(longley))
stopifnot(all.equal(cov2cor(Cv), cor(longley)))
}
