% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_rvce_logger}
\alias{get_rvce_logger}
\title{Create log4r Logger for package}
\usage{
get_rvce_logger(ps_logfile = "rvcetools.log", ps_level = "FATAL")
}
\arguments{
\item{ps_logfile}{name of the logfile}

\item{ps_level}{logger level}
}
\value{
rvcetools_logger
}
\description{
Create log4r Logger for package
}
\examples{
rvce_logger <- get_rvce_logger()
}
