% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVarCovarMix99.R
\name{parameter_varCovar_mix99}
\alias{parameter_varCovar_mix99}
\title{Parameter File for Mix99}
\usage{
parameter_varCovar_mix99(
  ps_input_file,
  ps_output_file,
  ps_raw_output_file = NULL,
  pn_ratio = NULL,
  pn_eps = NULL,
  pmat_weight = NULL,
  pn_digits = NULL,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_file}{input csv-file}

\item{ps_output_file}{output txt-file for the bended (co-)variances}

\item{ps_raw_output_file}{output txt-file for the raw (co-)variances}

\item{pn_ratio}{maximum ratio between largest and smallest eigenvalue, determines whether ratio method is used or not}

\item{pn_eps}{lower limit of smallest eigenvalue}

\item{pmat_weight}{weight matrix for weighted bending}

\item{pn_digits}{number of digits to be rounded to}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{log4r logger object}
}
\description{
Read csv-file of the variance component estimates and construct
a matrix with this estimates by parts. This matrix is checked if it is
positive definite or not. If this matrix is not positive definite, 2
functions may be called. If you put the paramter psOptionRatio = FALSE,
makePD2() is called . For the parameter psOptionRatio = TRUE the function
make_pd_rat_ev() is called. The difference is, that with make_pd_rat_ev() you
have the option to whish a maximum ratio between largest and smallest
eigenvalue.

The software mix99 need a paramter-File with the variance and covariance. So
this function is building this paramter-file.
}
\examples{
\dontrun{
# use test input csv-file
sInputFile <- system.file("extdata","VCE_results.csv", package = "rvcetools")
# bend matrices using Schaeffer method
parameter_varCovar_mix99(ps_input_file = sInputFile, ps_output_file = 'par_varCovar_mix99_schaeffer.txt')
# bend matrices using ratio method
parameter_varCovar_mix99(ps_input_file  = sInputFile, 
                         ps_output_file = 'par_varCovar_mix99_ratio.txt',
                         pn_ratio       = 100)
# unweighted bending with minimum eigenvalue
parameter_varCovar_mix99(ps_input_file  = sInputFile, 
                         ps_output_file = 'par_varCovar_mix99_unweighted.txt',
                         pn_eps         = 1e-4)
}
}
