% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositiveDefinit.R
\name{positivedefinit}
\alias{positivedefinit}
\title{Bending and Check or Transform Positive Definite Matrix}
\usage{
positivedefinit(ps_input_file, pn_ratio = NULL, pn_eps = NULL,
  pmat_weight = NULL, pn_digits = NULL, pb_log = FALSE,
  plogger = NULL)
}
\arguments{
\item{ps_input_file}{input csv-file}

\item{pn_ratio}{maximum ratio between largest and smallest eigenvalue, determines whether ratio method is used or not}

\item{pn_eps}{lower limit of smallest eigenvalue}

\item{pmat_weight}{weight matrix for weighted bending}

\item{pn_digits}{number of digits to which result will be rounded to}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{log4r logger object}
}
\description{
Read csv-file of the variance component estimates and
construct a matrix with this estimates by parts.
This matrix is checked if she is positive definite or not.
If this matrix is not positive definite, 2 functions may be called.
If pn_ratio == NULL, makePD2() is called, otherwise the function make_pd_rat_ev() is called.
The difference is, that with make_pd_rat_ev() you have the option to whish a maximum
ratio between largest and smallest eigenvalue.
}
\examples{
sInputFile <- system.file("extdata","VCE_results.csv", package = "rvcetools")
ResultPD <- positivedefinit(ps_input_file = sInputFile)
}
