% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVCE.R
\name{read_vce}
\alias{read_vce}
\title{Reading VCE Results From CSV File}
\usage{
read_vce(ps_input_file, pb_log = FALSE, plogger = NULL)
}
\arguments{
\item{ps_input_file}{name of the input file in csv-format}

\item{pb_log}{flag indicating whether logging should be done}

\item{plogger}{logger object from outside}
}
\description{
First attempt to read variance component estimation (VCE) results 
from a csv-file and storing them in tibble.
}
\examples{
# specify input file
sInputFile <- system.file("extdata","VCE_results.csv", package = "rvcetools")
tbl_vce <- read_vce(ps_input_file = sInputFile)
}
