% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{rvce_log_debug}
\alias{rvce_log_debug}
\title{Wrapper for log4r debug}
\usage{
rvce_log_debug(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r debug
}
\examples{
rvce_logger <- get_rvce_logger()
rvce_log_level(rvce_logger, 'DEBUG')
rvce_log_debug(rvce_logger)
}
