% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{rvce_log_error}
\alias{rvce_log_error}
\title{Wrapper for log4r error}
\usage{
rvce_log_error(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r error
}
\examples{
rvce_logger <- get_rvce_logger()
rvce_log_level(rvce_logger, 'ERROR')
rvce_log_error(rvce_logger)
}
