% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{rvce_log_fatal}
\alias{rvce_log_fatal}
\title{Wrapper for log4r fatal}
\usage{
rvce_log_fatal(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r fatal
}
\examples{
rvce_logger <- get_rvce_logger()
rvce_log_level(rvce_logger, 'FATAL')
rvce_log_fatal(rvce_logger)
}
