% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{rvce_log_info}
\alias{rvce_log_info}
\title{Wrapper for log4r info}
\usage{
rvce_log_info(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r info
}
\examples{
rvce_logger <- get_rvce_logger()
rvce_log_level(rvce_logger, 'INFO')
rvce_log_info(rvce_logger)
}
