% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{rvce_log_level}
\alias{rvce_log_level}
\title{Wrapper to set the level of a logger}
\usage{
rvce_log_level(
  plogger,
  ps_level = c("DEBUG", "INFO", "WARN", "ERROR", "FATAL")
)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_level}{new level of plogger}
}
\description{
Wrapper to set the level of a logger
}
\examples{
rvce_logger <- get_rvce_logger()
rvce_log_level(rvce_logger, 'INFO')
}
