% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{rvce_log_warn}
\alias{rvce_log_warn}
\title{Wrapper for log4r warn}
\usage{
rvce_log_warn(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r warn
}
\examples{
rvce_logger <- get_rvce_logger()
rvce_log_level(rvce_logger, 'WARN')
rvce_log_warn(rvce_logger)
}
