% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolvingErrorFromMiX99.R
\name{solving_errorFromMiX99_VCnotPD}
\alias{solving_errorFromMiX99_VCnotPD}
\title{Solving Message "VC matrix non positive definite" from MiX99}
\usage{
solving_errorFromMiX99_VCnotPD(
  ps_input_file,
  pn_ratio = NULL,
  pn_eps = NULL,
  pmat_weight = NULL,
  pn_digits = NULL,
  ps_output_file
)
}
\arguments{
\item{ps_input_file}{input file with VC in MiX99}

\item{pn_ratio}{specified ratio between largest and smallest eigenvalue}

\item{pn_eps}{smallest eigenvalue in weighted bending}

\item{pmat_weight}{weight matrix for weighted bending}

\item{pn_digits}{number of digits to be rounded to}

\item{ps_output_file}{solved file with VC for MiX99}
}
\description{
In the software MiX99 of Luke can the message "VC matrix non positive definite" poping up.
To solve this issue, the VC file non positive definite can be used as input and
a solution will be proposed as output file.
}
