% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Matrix.R
\name{build_matrix}
\alias{build_matrix}
\title{Convert Variance-Covariance Stored in Tibble to a List of matrices}
\usage{
build_matrix(psInputFile, pbLog = FALSE)
}
\arguments{
\item{psInputFile}{tibble containing all variance-covariance components}
}
\description{
Storing the variance and covariances in the tibble read from the input file
in a list of matrices.
}
\details{
---
title: Constructing multivariate estimates from analyses by parts in a matrix
date:  "`r Sys.Date()`"
---
}
\examples{
sInputFile <- system.file("extdata","VCE_results.csv", package = "rvcetools")
tbl_vce <- read_vce(ps_input_file = sInputFile)
l_mat <- build_matrix(psInputFile = tbl_vce)
}
