% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Matrix.R
\name{build_matrix}
\alias{build_matrix}
\title{Convert Variance-Covariance Stored in Tibble to a List of matrices}
\usage{
build_matrix(psInputFile, pbLog = FALSE)
}
\description{
Storing the variance and covariances in the tibble read from the input file
in a list of matrices.
}
\details{
---
title: Constructing multivariate estimates from analyses by parts in a matrix
date:  "`r Sys.Date()`"
---
}
