% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVarCovarMix99.R
\name{parameter_varCovar_mix99}
\alias{parameter_varCovar_mix99}
\title{Parameter File for Mix99}
\usage{
parameter_varCovar_mix99(psInputFile, psOptionRatio = FALSE,
  psRatio = 100, psOutputFile, pnDigits = NA)
}
\arguments{
\item{psInputFile}{input csv-file}

\item{psOptionRatio}{TRUE or FALSE}

\item{psRatio}{number}

\item{psOutputFile}{output txt-file}
}
\description{
Read csv-file of the variance component estimates and construct
a matrix with this estimates by parts. This matrix is checked if it is
positive definite or not. If this matrix is not positive definite, 2
functions may be called. If you put the paramter psOptionRatio = FALSE,
makePD2() is called . For the parameter psOptionRatio = TRUE the function
make_pd_rat_ev() is called. The difference is, that with make_pd_rat_ev() you
have the option to whish a maximum ratio between largest and smallest
eigenvalue.

The software mix99 need a paramter-File with the variance and covariance. So
this function is building this paramter-file.
}
\examples{
\dontrun{
sInputFile <- system.file("extdata","VCE_results.csv", package = "rvcetools")
parameter_varCovar_mix99(psInputFile = sInputFile, psOutputFile = 'par_varCovar_mix99.txt')
}
}
