% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositiveDefinit.R
\name{positivedefinit}
\alias{positivedefinit}
\title{Bending and Check or Transform Positive Definite Matrix}
\usage{
positivedefinit(psInputFile, psOptionRatio = FALSE, psRatio = 100,
  pnDigits = NA)
}
\arguments{
\item{psInputFile}{input csv-file}

\item{psOptionRatio}{TRUE or FALSE (default) to indicate whether bending should be done based on ratio of largest to smallest eigenvalue}

\item{psRatio}{maximum ratio between largest and smallest eigenvalue. Only relevant, if psOptionRatio == TRUE}
}
\description{
Read csv-file of the variance component estimates and
construct a matrix with this estimates by parts.
This matrix is checked if she is positive definite or not.
If this matrix is not positive definite, 2 functions may be called.
If you put the paramter psOptionRatio = FALSE, makePD2() is called .
For the parameter psOptionRatio = TRUE the function make_pd_rat_ev() is called.
The difference is, that with make_pd_rat_ev() you have the option to whish a maximum
ratio between largest and smallest eigenvalue.
}
